+++ exp/services/ledgerexporter/internal/datastore.go
@@ -3,14 +3,9 @@ package exporter
 import (
 	"context"
 	"io"
-	"os"
-	"path"
 	"strings"
 
-	"google.golang.org/api/googleapi"
-
 	"cloud.google.com/go/storage"
-	log "github.com/sirupsen/logrus"
 	"github.com/stellar/go/support/errors"
 	"github.com/stellar/go/support/url"
 	"google.golang.org/api/option"
@@ -24,13 +19,6 @@ type DataStore interface {
 	Exists(path string) (bool, error)
 	Size(path string) (int64, error)
 	Close() error
-}
-
-// GCSDataStore implements DataStore for GCS
-type GCSDataStore struct {
-	client *storage.Client
-	bucket *storage.BucketHandle
-	prefix string
 }
 
 // NewDataStore creates a new DataStore based on the destination URL.
@@ -51,10 +39,7 @@ func NewDataStore(destinationURL string) (DataStore, error) {
 	bucketName := parsed.Host
 	prefix := pth
 
-	log.WithFields(log.Fields{
-		"bucket": bucketName,
-		"prefix": prefix,
-	}).Debug("gcs: making backend")
+	logger.Infof("creating GCS client for bucket: %s, prefix: %s", bucketName, prefix)
 
 	var options []option.ClientOption
 	client, err := storage.NewClient(context.Background(), options...)
@@ -69,78 +54,4 @@ func NewDataStore(destinationURL string) (DataStore, error) {
 	}
 
 	return &GCSDataStore{client: client, bucket: bucket, prefix: prefix}, nil
-}
-
-// GetFile retrieves a file from the GCS bucket.
-func (b *GCSDataStore) GetFile(filePath string) (io.ReadCloser, error) {
-	filePath = path.Join(b.prefix, filePath)
-	log.WithField("path", filePath).Trace("gcs: get file")
-
-	r, err := b.bucket.Object(filePath).NewReader(context.Background())
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to get file: %s", filePath)
-	}
-
-	return r, nil
-}
-
-// PutFileIfNotExists uploads a file to GCS only if it doesn't already exist.
-func (b *GCSDataStore) PutFileIfNotExists(pth string, in io.ReadCloser) error {
-	err := b.putFile(pth, in, &storage.Conditions{DoesNotExist: true})
-	if e, ok := err.(*googleapi.Error); ok {
-		if e.Code == 412 {
-			log.WithField("path", pth).Info("File already exists in the bucket, skipping upload.")
-			return nil // Treat as success since the file is already present
-		}
-	}
-	return err
-}
-
-// PutFile uploads a file to GCS
-func (b *GCSDataStore) PutFile(pth string, in io.ReadCloser) error {
-	return b.putFile(pth, in, nil) // No conditions for regular PutFile
-}
-
-// Size retrieves the size of a file in the GCS bucket.
-func (b *GCSDataStore) Size(pth string) (int64, error) {
-	pth = path.Join(b.prefix, pth)
-	log.WithField("path", pth).Trace("gcs: get size")
-	attrs, err := b.bucket.Object(pth).Attrs(context.Background())
-	if err == storage.ErrObjectNotExist {
-		err = os.ErrNotExist
-	}
-	if err != nil {
-		return 0, err
-	}
-	return attrs.Size, nil
-}
-
-// Exists checks if a file exists in the GCS bucket.
-func (b *GCSDataStore) Exists(pth string) (bool, error) {
-	log.WithField("path", path.Join(b.prefix, pth)).Trace("gcs: check exists")
-	_, err := b.Size(pth)
-	return err == nil, err
-}
-
-// Close closes the GCS client connection.
-func (b *GCSDataStore) Close() error {
-	log.Trace("gcs: close")
-	return b.client.Close()
-}
-
-func (b *GCSDataStore) putFile(filePath string, in io.ReadCloser, conditions *storage.Conditions) error {
-	filePath = path.Join(b.prefix, filePath)
-	log.WithField("path", filePath).Trace("gcs: put file")
-
-	o := b.bucket.Object(filePath)
-	if conditions != nil {
-		o = o.If(*conditions)
-	}
-	w := o.NewWriter(context.Background())
-	if _, err := io.Copy(w, in); err != nil {
-		in.Close()
-		return errors.Wrapf(err, "failed to put file: %s", filePath)
-	}
-
-	return w.Close()
 }
